refactor: extract NetworkProvider
